Replaced interface MalwareClientStatusMapper with implementation
