--- cds-feature-attachments/src/main/java/com/sap/cds/feature/attachments/service/malware/client/mapper/DefaultMalwareClientStatusMapper.java
@@ -1,20 +1,0 @@
-/**************************************************************************
- * (C) 2019-2024 SAP SE or an SAP affiliate company. All rights reserved. *
- **************************************************************************/
-package com.sap.cds.feature.attachments.service.malware.client.mapper;
-
-import com.sap.cds.feature.attachments.generated.cds4j.sap.attachments.StatusCode;
-import com.sap.cds.feature.attachments.service.malware.client.model.MalwareScanResultStatus;
-
-public class DefaultMalwareClientStatusMapper implements MalwareClientStatusMapper {
-
-	@Override
-	public String mapStatus(MalwareScanResultStatus status) {
-		return switch (status) {
-			case CLEAN, NO_SCANNER -> StatusCode.CLEAN;
-			case INFECTED, ENCRYPTED -> StatusCode.INFECTED;
-			default -> StatusCode.FAILED;
-		};
-	}
-
-}
