+++ cds-feature-attachments/src/main/java/com/sap/cds/feature/attachments/service/malware/client/mapper/MalwareClientStatusMapper.java
@@ -3,10 +3,17 @@
  **************************************************************************/
 package com.sap.cds.feature.attachments.service.malware.client.mapper;
 
+import com.sap.cds.feature.attachments.generated.cds4j.sap.attachments.StatusCode;
 import com.sap.cds.feature.attachments.service.malware.client.model.MalwareScanResultStatus;
 
-public interface MalwareClientStatusMapper {
+public class MalwareClientStatusMapper {
 
-	String mapStatus(MalwareScanResultStatus status);
+	public String mapStatus(MalwareScanResultStatus status) {
+		return switch (status) {
+			case CLEAN, NO_SCANNER -> StatusCode.CLEAN;
+			case INFECTED, ENCRYPTED -> StatusCode.INFECTED;
+			default -> StatusCode.FAILED;
+		};
+	}
 
 }
