+++ cds-feature-attachments/src/test/java/com/sap/cds/feature/attachments/service/malware/client/mapper/DefaultMalwareClientStatusMapperTest.java
@@ -13,11 +13,11 @@ import com.sap.cds.feature.attachments.service.malware.client.model.MalwareScanR
 
 class DefaultMalwareClientStatusMapperTest {
 
-	private DefaultMalwareClientStatusMapper cut;
+	private MalwareClientStatusMapper cut;
 
 	@BeforeEach
 	void setup() {
-		cut = new DefaultMalwareClientStatusMapper();
+		cut = new MalwareClientStatusMapper();
 	}
 
 	@Test
